Refactor watch/cancellable actions
